refactor(call): extract _SignalingModule extension from CallBloc (#1017)
Move all signaling-related methods (~880 lines) from CallBloc into a
dedicated part file (signaling_module.dart) as a private extension.

Includes:
- _reconnectInitiated / _disconnectInitiated
- _onSignalingClientEvent and all sub-handlers
- _onHandshakeSignalingEventState
- _onCallSignalingEvent and all 14 __onCallSignalingEvent* handlers
- _onSignalingStateHandshake / _onSignalingEvent / _onSignalingError / _onSignalingDisconnect

on<> registrations in CallBloc use lambda wrappers to call extension methods.
